Implement basic HTTP proxy functionality
This commit adds ability to use this tool as a basic HTTP proxy.
It takes bytes from the TCP connection, forwards them to the destination
server, reads the response and forwards it back to the client.
It uses the value from the Host header in the request to determine the
address of the destination server.
Request is limited to 512B and response is limited to 64KB.
Default bind address is 127.0.0.1:7878, but this will be configurable in
cli argument in the future.